requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.16
django-cors-headers==3.3.0
django-grappelli==2.13.2
django-import-export==3.0.1
django-jet==1.0.8
django-mailjet==0.3.1
django-sms==0.5.0
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
future==0.18.2
google-api-python-client==1.4.1
httplib2==0.14.0
idna==3.4
jdcal==1.4.1
mailjet==1.4.0
mailjet-rest==1.3.4
MarkupPy==1.14
oauth2client==4.1.3
odfpy==1.4.0
openpyxl==3.0.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-decouple==3.3
pytz==2019.3
PyYAML==5.1.2
requests==2.28.1
rsa==4.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
tablib==3.3.0
typing-extensions==4.4.0
uritemplate==3.0.0
urllib3==1.26.13
xlrd==1.2.0
xlwt==1.3.0
